Authentication - Moved to resource server based authentication - Created a seperate authentication server project - Created authentication service and controller - Created a custom username password authentication filter - Moved JwtAuthenticationConverter and UserDetailsService to their own files

+++ src/main/kotlin/com/kio/configuration/properties/OAuth2ConfigurationProperties.kt
@@ -8,6 +8,6 @@ import org.springframework.boot.context.properties.ConstructorBinding
 data class OAuth2ConfigurationProperties(
     val id: String,
     val secret: String,
-    val refreshTokenValidityTime: Int,
-    val accessTokenValidityTime: Int,
+    val refreshTokenValidityTime: Long,
+    val accessTokenValidityTime: Long,
 )
